+++ DotaHeroes/API/Statistics/ResistanceStatistics.cs
@@ -6,13 +6,13 @@ namespace DotaHeroes.API.Statistics
 {
     public class ResistanceStatistics
     {
-        public decimal BaseMagicResistance { get; set; }
+        public double BaseMagicResistance { get; set; }
 
-        public decimal BaseEffectResistance { get; set; }
+        public double BaseEffectResistance { get; set; }
 
         public List<IResistanceModifier> ResistanceModifiers { get; set; }
 
-        public const decimal BaseResistance = 25;
+        public const double BaseResistance = 25;
 
         public ResistanceStatistics()
         {
@@ -21,43 +21,43 @@ namespace DotaHeroes.API.Statistics
             ResistanceModifiers = new List<IResistanceModifier>();
         }
 
-        public ResistanceStatistics(decimal magicResistance, decimal baseEffectResistance)
+        public ResistanceStatistics(double magicResistance, double baseEffectResistance)
         {
             BaseMagicResistance = magicResistance;
             BaseEffectResistance = baseEffectResistance;
             ResistanceModifiers = new List<IResistanceModifier>();
         }
 
-        public decimal GetMagicalDamage(decimal damage, decimal intelligence)
+        public double GetMagicalDamage(double damage, double intelligence)
         {
-            decimal percent = (damage / 100m);
+            double percent = (damage / 100);
             return damage - percent * GetMagicResistance(intelligence);
         }
 
-        public decimal GetEffectResistance()
+        public double GetEffectResistance()
         {
-            decimal baseEffectResistance = (BaseEffectResistance / 100);
-            decimal effectResistanceFromModifiers = GetEffectResistanceFromModifiers();
+            double baseEffectResistance = (BaseEffectResistance / 100);
+            double effectResistanceFromModifiers = GetEffectResistanceFromModifiers();
 
             return (1 - ((1 - baseEffectResistance) * effectResistanceFromModifiers)) * 100;
         }
 
-        public decimal GetEffectDuration(decimal originalDuration)
+        public double GetEffectDuration(double originalDuration)
         {
             return originalDuration * ((100 - GetEffectResistance()) / 100);
         }
 
-        public decimal GetMagicResistance(decimal intelligence)
+        public double GetMagicResistance(double intelligence)
         {
-            decimal baseMagicResistance = (BaseMagicResistance / 100);
-            decimal magicResistanceFromModifiers = GetMagicResistanceFromModifiers();
+            double baseMagicResistance = (BaseMagicResistance / 100);
+            double magicResistanceFromModifiers = GetMagicResistanceFromModifiers();
 
             return ((1 - ((1 - baseMagicResistance) * magicResistanceFromModifiers)) + ((intelligence / 10) / 100)) * 100;
         }
 
-        private decimal GetEffectResistanceFromModifiers()
+        private double GetEffectResistanceFromModifiers()
         {
-            decimal result = 1;
+            double result = 1;
 
             foreach (var modifier in ResistanceModifiers)
             {
@@ -67,9 +67,9 @@ namespace DotaHeroes.API.Statistics
             return result;
         }
 
-        private decimal GetMagicResistanceFromModifiers()
+        private double GetMagicResistanceFromModifiers()
         {
-            decimal result = 1;
+            double result = 1;
 
             foreach (var modifier in ResistanceModifiers)
             {
@@ -84,7 +84,7 @@ namespace DotaHeroes.API.Statistics
             return $"Effect resistance: {GetEffectResistance()}";
         }
 
-        public string ToString(decimal intelligence)
+        public string ToString(double intelligence)
         {
             return $"Magic resistance: {Math.Round(GetMagicResistance(intelligence), 1)} Effect resistance: {Math.Round(GetEffectResistance(), 1)}";
         }
